requirements.txt
appnope==0.1.3
asttokens==2.0.8
backcall==0.2.0
cloudpickle==2.2.0
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
drake==1.9.0
entrypoints==0.4
executing==1.1.1
fonttools==4.38.0
gradescope-utils==0.5.0
ipykernel==6.16.1
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.3
jupyter_core==4.11.2
kiwisolver==1.4.4
manipulation==2022.9.23
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
meshcat==0.3.2
mpld3==0.5.8
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydot==1.4.2
Pygments==2.13.0
pyngrok==5.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==24.0.1
six==1.16.0
stack-data==0.5.1
timeout-decorator==0.5.0
tornado==6.2
traitlets==5.5.0
u-msgpack-python==2.7.1
wcwidth==0.2.5
